LPD-77438 Decouple wizard and context
