+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/js/pages/import/NewImport.tsx
@@ -5,12 +5,21 @@
 
 import ClayButton from '@clayui/button';
 import ClayIcon from '@clayui/icon';
-import React from 'react';
+import React, {createContext, useContext} from 'react';
 
 import {Wizard, WizardStep} from '../../components/Wizard';
 import DataSelectionStep from './steps/DataSelectionStep';
 import FileSelectionStep from './steps/FileSelectionStep';
 import SettingsStep from './steps/SettingsStep';
+
+export const WizardContext = createContext({
+	groupId: -1,
+	isCompanyGroup: false,
+});
+
+export function useWizard() {
+	return useContext(WizardContext);
+}
 
 export function NewImport({
 	backURL,
@@ -22,52 +31,50 @@ export function NewImport({
 	isCompanyGroup: boolean;
 }) {
 	return (
-		<Wizard
-			backURL={backURL}
-			groupId={groupId}
-			isCompanyGroup={isCompanyGroup}
-		>
-			<WizardStep
-				description={Liferay.Language.get(
-					'name-your-import-process-and-upload-your-file'
-				)}
-				initialValues={{
-					fileSelector: undefined,
-				}}
-				title={Liferay.Language.get('setup')}
-			>
-				<FileSelectionStep />
-			</WizardStep>
+		<WizardContext.Provider value={{groupId, isCompanyGroup}}>
+			<Wizard backURL={backURL}>
+				<WizardStep
+					description={Liferay.Language.get(
+						'name-your-import-process-and-upload-your-file'
+					)}
+					initialValues={{
+						fileSelector: undefined,
+					}}
+					title={Liferay.Language.get('setup')}
+				>
+					<FileSelectionStep />
+				</WizardStep>
 
-			<WizardStep
-				description={Liferay.Language.get(
-					'select-the-data-from-your-file-that-you-would-like-to-import'
-				)}
-				title={Liferay.Language.get('data-selection')}
-			>
-				<DataSelectionStep />
-			</WizardStep>
+				<WizardStep
+					description={Liferay.Language.get(
+						'select-the-data-from-your-file-that-you-would-like-to-import'
+					)}
+					title={Liferay.Language.get('data-selection')}
+				>
+					<DataSelectionStep />
+				</WizardStep>
 
-			<WizardStep
-				actionButton={
-					<ClayButton type="submit">
-						<span className="inline-item inline-item-before">
-							<ClayIcon className="mr-1" symbol="import" />
-						</span>
+				<WizardStep
+					actionButton={
+						<ClayButton type="submit">
+							<span className="inline-item inline-item-before">
+								<ClayIcon className="mr-1" symbol="import" />
+							</span>
 
-						{Liferay.Language.get('import')}
-					</ClayButton>
-				}
-				description={Liferay.Language.get(
-					'set-up-your-import-configuration'
-				)}
-				onSubmit={async () => {
-					alert('Import started!');
-				}}
-				title={Liferay.Language.get('settings')}
-			>
-				<SettingsStep />
-			</WizardStep>
-		</Wizard>
+							{Liferay.Language.get('import')}
+						</ClayButton>
+					}
+					description={Liferay.Language.get(
+						'set-up-your-import-configuration'
+					)}
+					onSubmit={async () => {
+						alert('Import started!');
+					}}
+					title={Liferay.Language.get('settings')}
+				>
+					<SettingsStep />
+				</WizardStep>
+			</Wizard>
+		</WizardContext.Provider>
 	);
 }
